refactor(channels): add maintainability guardrails
