+++ .husky/pre-commit
@@ -46,6 +46,18 @@ if ! node scripts/check-filename-casing.js --staged 2>/dev/null; then
   exit 1
 fi
 
+# Keep source modules small and prevent ownership entrypoints from becoming barrels.
+if ! node scripts/check-source-file-size.js 2>/dev/null; then
+  echo ""
+  echo "❌ Source file size limit failed — run 'bun run check:file-size' for details."
+  exit 1
+fi
+if ! node scripts/check-module-ownership.js 2>/dev/null; then
+  echo ""
+  echo "❌ Module ownership failed — run 'bun run check:module-ownership' for details."
+  exit 1
+fi
+
 # Check staged skill metadata for required name headers
 if ! node scripts/check-skill-frontmatter.js --staged; then
   echo ""
