+++ package.json
@@ -95,6 +95,8 @@
     "check:boundaries": "node scripts/check-layer-boundaries.js",
     "check:exported-functions": "node scripts/check-exported-functions.js",
     "check:filename-casing": "node scripts/check-filename-casing.js",
+    "check:file-size": "node scripts/check-source-file-size.js",
+    "check:module-ownership": "node scripts/check-module-ownership.js",
     "check:test-mock-isolation": "bun run scripts/check-test-mock-isolation.js",
     "check:test-coverage": "node scripts/check-test-coverage.cjs",
     "check:skill-frontmatter": "node scripts/check-skill-frontmatter.js",
