+++ scripts/check.js
@@ -17,6 +17,8 @@ const checks = [
   { name: "layer boundaries", script: ["check:boundaries"] },
   { name: "exported function style", script: ["check:exported-functions"] },
   { name: "filename casing", script: ["check:filename-casing"] },
+  { name: "source file size", script: ["check:file-size"] },
+  { name: "module ownership", script: ["check:module-ownership"] },
   { name: "test mock isolation", script: ["check:test-mock-isolation"] },
   { name: "test coverage", script: ["check:test-coverage"] },
   { name: "skill frontmatter", script: ["check:skill-frontmatter"] },
