+++ scripts/source-file-size-baseline.json
@@ -1,0 +1,65 @@
+{
+  "src/agent/client-skills.test.ts": 1196,
+  "src/agent/memory-git.ts": 2324,
+  "src/agent/subagents/manager.ts": 1577,
+  "src/backend/local-backend.test.ts": 2589,
+  "src/backend/local/local-backend.ts": 1058,
+  "src/backend/local/local-store.ts": 3629,
+  "src/backend/pi-stream-adapter.test.ts": 1442,
+  "src/channels/discord/adapter.ts": 1279,
+  "src/channels/message-channel.test.ts": 1639,
+  "src/channels/registry.test.ts": 1936,
+  "src/channels/registry.ts": 2872,
+  "src/channels/schema-config.test.ts": 1041,
+  "src/channels/service.ts": 2096,
+  "src/channels/signal/adapter.ts": 1295,
+  "src/channels/signal/setup.ts": 1031,
+  "src/channels/telegram-adapter.test.ts": 2677,
+  "src/channels/telegram/adapter.ts": 1574,
+  "src/cli/app/AppCoordinator.tsx": 5085,
+  "src/cli/app/AppView.tsx": 1724,
+  "src/cli/app/use-approval-flow.ts": 1163,
+  "src/cli/app/use-configuration-handlers.ts": 1384,
+  "src/cli/app/use-conversation-loop.ts": 2935,
+  "src/cli/app/use-submit-handler.ts": 3768,
+  "src/cli/components/AgentSelector.tsx": 1270,
+  "src/cli/components/InputRich.tsx": 2219,
+  "src/cli/components/ModelSelector.tsx": 1299,
+  "src/cli/components/ProviderSelector.tsx": 1676,
+  "src/cli/components/ToolCallMessageRich.tsx": 1109,
+  "src/cli/helpers/accumulator.ts": 1596,
+  "src/cli/helpers/reflection-transcript.ts": 2080,
+  "src/cli/helpers/stream.ts": 1046,
+  "src/cli/mods/local-mod-loader.test.ts": 1043,
+  "src/cli/reflection-transcript.test.ts": 1118,
+  "src/cli/subcommands/skills.ts": 1264,
+  "src/headless.ts": 5238,
+  "src/hooks/integration.test.ts": 1147,
+  "src/index.ts": 2854,
+  "src/mods/learning-harness.ts": 2434,
+  "src/mods/mod-engine.test.ts": 2178,
+  "src/mods/mod-engine.ts": 1919,
+  "src/mods/package-installer.test.ts": 1248,
+  "src/mods/package-installer.ts": 1201,
+  "src/mods/package-registry.ts": 1033,
+  "src/permissions/read-only-shell.test.ts": 1303,
+  "src/permissions/read-only-shell.ts": 2009,
+  "src/providers/chatgpt-usage-service.ts": 1115,
+  "src/settings-manager.test.ts": 1669,
+  "src/settings-manager.ts": 2113,
+  "src/tools/impl/enter-worktree.ts": 1265,
+  "src/tools/impl/message-channel.ts": 1208,
+  "src/tools/manager.ts": 3293,
+  "src/tools/tool-execution-context.test.ts": 1422,
+  "src/types/protocol_v2.ts": 2931,
+  "src/websocket/listen-client-concurrency.test.ts": 3110,
+  "src/websocket/listen-client-protocol.test.ts": 6942,
+  "src/websocket/listener/commands/channels.ts": 1390,
+  "src/websocket/listener/commands/memory.ts": 1114,
+  "src/websocket/listener/file-commands.ts": 1030,
+  "src/websocket/listener/lifecycle.ts": 1768,
+  "src/websocket/listener/message-router.ts": 1011,
+  "src/websocket/listener/protocol-inbound.ts": 2279,
+  "src/websocket/listener/protocol-outbound.ts": 1295,
+  "src/websocket/listener/turn.ts": 1495
+}
