+++ src/channels/slack/adapter.test.ts
@@ -1,303 +1,541 @@
-import { describe, expect, test } from "bun:test";
+import { expect, mock, test } from "bun:test";
+import { mkdtemp, writeFile } from "node:fs/promises";
+import { tmpdir } from "node:os";
+import { join } from "node:path";
 import {
-  buildSlackDebounceKey,
-  buildTopLevelSlackConversationKey,
-  createAgentThreadTracker,
-  resolveSlackInboundDebounceMs,
-} from "@/channels/slack/adapter";
+  createSlackAdapter,
+  FakeSlackApp,
+  FakeSlackWriteClient,
+  fetchMock,
+  installSlackAdapterTestHooks,
+  resolveSlackAccountDisplayName,
+  slackAccountDefaults,
+} from "./adapter-test-harness";
+import { buildSlackModelPickerBlocks } from "./model-picker-blocks";
 
-const ACCOUNT = "acct-1";
+installSlackAdapterTestHooks();
 
-describe("buildSlackDebounceKey", () => {
-  test("DMs are channel-scoped (two messages from same user in same DM share a key)", () => {
-    const k1 = buildSlackDebounceKey(
-      { channel: "D123", ts: "1700000000.000001", user: "U1" },
-      ACCOUNT,
-    );
-    const k2 = buildSlackDebounceKey(
-      { channel: "D123", ts: "1700000000.000002", user: "U1" },
-      ACCOUNT,
-    );
-    expect(k1).not.toBeNull();
-    expect(k1).toBe(k2);
+test("slack adapter start does not re-run bolt init", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
   });
 
-  test("DMs from different users produce different keys", () => {
-    const k1 = buildSlackDebounceKey(
-      { channel: "D123", ts: "1.000001", user: "U1" },
-      ACCOUNT,
-    );
-    const k2 = buildSlackDebounceKey(
-      { channel: "D123", ts: "1.000002", user: "U2" },
-      ACCOUNT,
-    );
-    expect(k1).not.toBe(k2);
-  });
+  await adapter.start();
 
-  test("thread replies are thread-scoped", () => {
-    const k1 = buildSlackDebounceKey(
-      {
-        channel: "C123",
-        ts: "1700000000.000001",
-        thread_ts: "1699999999.000000",
-        user: "U1",
-      },
-      ACCOUNT,
-    );
-    const k2 = buildSlackDebounceKey(
-      {
-        channel: "C123",
-        ts: "1700000000.000002",
-        thread_ts: "1699999999.000000",
-        user: "U1",
-      },
-      ACCOUNT,
-    );
-    expect(k1).toBe(k2);
-  });
-
-  test("probable thread reply (parent_user_id, no thread_ts) uses maybe-thread key", () => {
-    const key = buildSlackDebounceKey(
-      {
-        channel: "C123",
-        ts: "1700000000.000001",
-        parent_user_id: "U_parent",
-        user: "U1",
-      },
-      ACCOUNT,
-    );
-    expect(key).toContain("maybe-thread:1700000000.000001");
-  });
-
-  test("top-level channel posts are message-ts-scoped (different posts do not merge)", () => {
-    const k1 = buildSlackDebounceKey(
-      { channel: "C123", ts: "1700000000.000001", user: "U1" },
-      ACCOUNT,
-    );
-    const k2 = buildSlackDebounceKey(
-      { channel: "C123", ts: "1700000000.000002", user: "U1" },
-      ACCOUNT,
-    );
-    expect(k1).not.toBe(k2);
-  });
-
-  test("missing user (and no bot_id) → null", () => {
-    const key = buildSlackDebounceKey({ channel: "C123", ts: "1.0" }, ACCOUNT);
-    expect(key).toBeNull();
-  });
-
-  test("bot_id is accepted as sender when user is missing", () => {
-    const key = buildSlackDebounceKey(
-      { channel: "C123", ts: "1.0", bot_id: "B42" },
-      ACCOUNT,
-    );
-    expect(key).not.toBeNull();
-    expect(key).toContain("B42");
-  });
-
-  test("event_ts is used when ts is missing", () => {
-    const key = buildSlackDebounceKey(
-      { channel: "C123", event_ts: "1.0", user: "U1" },
-      ACCOUNT,
-    );
-    expect(key).not.toBeNull();
-    expect(key).toContain("1.0");
-  });
-
-  test("accountId scopes the key (different accounts → different keys)", () => {
-    const k1 = buildSlackDebounceKey(
-      { channel: "D123", ts: "1.0", user: "U1" },
-      "acct-a",
-    );
-    const k2 = buildSlackDebounceKey(
-      { channel: "D123", ts: "1.0", user: "U1" },
-      "acct-b",
-    );
-    expect(k1).not.toBe(k2);
-  });
+  const app = FakeSlackApp.instances[0];
+  expect(app).toBeDefined();
+  expect(app?.init).not.toHaveBeenCalled();
+  expect(app?.start).toHaveBeenCalledTimes(1);
 });
 
-describe("buildTopLevelSlackConversationKey", () => {
-  test("returns a key for top-level channel posts", () => {
-    const key = buildTopLevelSlackConversationKey(
-      { channel: "C123", ts: "1.0", user: "U1" },
-      ACCOUNT,
-    );
-    expect(key).not.toBeNull();
+test("slack adapter forwards native channel slash commands as channel slash input", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
   });
+  const messages: unknown[] = [];
+  adapter.onMessage = async (message) => {
+    messages.push(message);
+  };
 
-  test("returns null for DMs", () => {
-    const key = buildTopLevelSlackConversationKey(
-      { channel: "D123", ts: "1.0", user: "U1" },
-      ACCOUNT,
-    );
-    expect(key).toBeNull();
-  });
-
-  test("returns null for thread replies", () => {
-    const key = buildTopLevelSlackConversationKey(
-      {
-        channel: "C123",
-        ts: "1.0",
-        thread_ts: "0.9",
-        user: "U1",
-      },
-      ACCOUNT,
-    );
-    expect(key).toBeNull();
-  });
-
-  test("returns null when parent_user_id is set (maybe-thread case)", () => {
-    const key = buildTopLevelSlackConversationKey(
-      {
-        channel: "C123",
-        ts: "1.0",
-        parent_user_id: "U_parent",
-        user: "U1",
-      },
-      ACCOUNT,
-    );
-    expect(key).toBeNull();
-  });
-
-  test("returns null when sender is missing", () => {
-    const key = buildTopLevelSlackConversationKey(
-      { channel: "C123", ts: "1.0" },
-      ACCOUNT,
-    );
-    expect(key).toBeNull();
-  });
-});
-
-describe("resolveSlackInboundDebounceMs", () => {
-  const originalEnv = process.env.LETTA_SLACK_INBOUND_DEBOUNCE_MS;
-
-  function clearEnv() {
-    delete process.env.LETTA_SLACK_INBOUND_DEBOUNCE_MS;
+  await adapter.start();
+  const app = FakeSlackApp.instances[0];
+  const handler = app?.commandHandlers.get("/cancel");
+  if (!handler) {
+    throw new Error("Expected /cancel command handler");
   }
 
-  function restoreEnv() {
-    if (originalEnv === undefined) clearEnv();
-    else process.env.LETTA_SLACK_INBOUND_DEBOUNCE_MS = originalEnv;
+  const ack = mock(async () => {});
+  await handler({
+    command: {
+      command: "/cancel",
+      text: "",
+      user_id: "U123",
+      user_name: "Alice",
+      channel_id: "C123",
+      channel_name: "eng",
+      trigger_id: "trigger-1",
+    },
+    ack,
+  });
+
+  const modelHandler = app?.commandHandlers.get("/model");
+  if (!modelHandler) {
+    throw new Error("Expected /model command handler");
   }
 
-  test("defaults to 0 when no env var and no config value", () => {
-    clearEnv();
-    expect(resolveSlackInboundDebounceMs({})).toBe(0);
-    restoreEnv();
+  await modelHandler({
+    command: {
+      command: "/model",
+      text: "openai/gpt-5",
+      user_id: "U123",
+      user_name: "Alice",
+      channel_id: "C123",
+      channel_name: "eng",
+      trigger_id: "trigger-2",
+    },
+    ack,
   });
 
-  test("returns config value when env is unset", () => {
-    clearEnv();
-    expect(resolveSlackInboundDebounceMs({ inboundDebounceMs: 1500 })).toBe(
-      1500,
-    );
-    restoreEnv();
+  expect(ack).toHaveBeenCalledTimes(2);
+  expect(messages).toHaveLength(2);
+  expect(messages[0]).toMatchObject({
+    channel: "slack",
+    accountId: "slack-test-account",
+    chatId: "C123",
+    senderId: "U123",
+    senderName: "Alice",
+    chatLabel: "eng",
+    text: "/cancel",
+    messageId: "trigger-1",
+    threadId: null,
+    chatType: "channel",
   });
-
-  test("config value of 0 is respected when env is unset", () => {
-    clearEnv();
-    expect(resolveSlackInboundDebounceMs({ inboundDebounceMs: 0 })).toBe(0);
-    restoreEnv();
-  });
-
-  test("env var overrides config", () => {
-    clearEnv();
-    process.env.LETTA_SLACK_INBOUND_DEBOUNCE_MS = "2500";
-    expect(resolveSlackInboundDebounceMs({ inboundDebounceMs: 1500 })).toBe(
-      2500,
-    );
-    restoreEnv();
-  });
-
-  test("env var of 0 is respected (disables debounce even with config set)", () => {
-    clearEnv();
-    process.env.LETTA_SLACK_INBOUND_DEBOUNCE_MS = "0";
-    expect(resolveSlackInboundDebounceMs({ inboundDebounceMs: 1500 })).toBe(0);
-    restoreEnv();
-  });
-
-  test("invalid env var falls back to config", () => {
-    clearEnv();
-    process.env.LETTA_SLACK_INBOUND_DEBOUNCE_MS = "not-a-number";
-    expect(resolveSlackInboundDebounceMs({ inboundDebounceMs: 800 })).toBe(800);
-    restoreEnv();
-  });
-
-  test("invalid env var falls back to disabled when config is absent", () => {
-    clearEnv();
-    process.env.LETTA_SLACK_INBOUND_DEBOUNCE_MS = "not-a-number";
-    expect(resolveSlackInboundDebounceMs({})).toBe(0);
-    restoreEnv();
-  });
-
-  test("negative env var falls back to config or disabled", () => {
-    clearEnv();
-    process.env.LETTA_SLACK_INBOUND_DEBOUNCE_MS = "-1";
-    expect(resolveSlackInboundDebounceMs({ inboundDebounceMs: 800 })).toBe(800);
-    expect(resolveSlackInboundDebounceMs({})).toBe(0);
-    restoreEnv();
-  });
-
-  test("invalid or negative config value falls back to disabled", () => {
-    clearEnv();
-    expect(resolveSlackInboundDebounceMs({ inboundDebounceMs: -500 })).toBe(0);
-    expect(
-      resolveSlackInboundDebounceMs({ inboundDebounceMs: Number.NaN }),
-    ).toBe(0);
-    restoreEnv();
-  });
-
-  test("empty-string env var is treated as unset", () => {
-    clearEnv();
-    process.env.LETTA_SLACK_INBOUND_DEBOUNCE_MS = "";
-    expect(resolveSlackInboundDebounceMs({ inboundDebounceMs: 1234 })).toBe(
-      1234,
-    );
-    restoreEnv();
+  expect(messages[1]).toMatchObject({
+    channel: "slack",
+    accountId: "slack-test-account",
+    chatId: "C123",
+    senderId: "U123",
+    senderName: "Alice",
+    chatLabel: "eng",
+    text: "/model openai/gpt-5",
+    messageId: "trigger-2",
+    threadId: null,
+    chatType: "channel",
   });
 });
 
-describe("createAgentThreadTracker", () => {
-  test("remembers and checks channel threads", () => {
-    const tracker = createAgentThreadTracker();
-    expect(tracker.has("C123", "1234567890.123456")).toBe(false);
-    tracker.remember("C123", "1234567890.123456");
-    expect(tracker.has("C123", "1234567890.123456")).toBe(true);
-    expect(tracker.has("C123", "9999999999.999999")).toBe(false);
+test("slack adapter forwards model picker selections as /model commands", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+  const messages: unknown[] = [];
+  adapter.onMessage = async (message) => {
+    messages.push(message);
+  };
+
+  await adapter.start();
+  const app = FakeSlackApp.instances[0];
+  const handler = app?.actionHandlers.get("letta_channel_model_select");
+  if (!handler) {
+    throw new Error("Expected model select action handler");
+  }
+
+  const ack = mock(async () => {});
+  await handler({
+    body: {
+      user: { id: "U123", name: "Alice", team_id: "T123" },
+      channel: { id: "C123", name: "eng" },
+      container: {
+        channel_id: "C123",
+        message_ts: "1712800000.000100",
+        thread_ts: "1712800000.000200",
+      },
+      message: { ts: "1712800000.000100", thread_ts: "1712800000.000200" },
+    },
+    action: {
+      action_id: "letta_channel_model_select",
+      action_ts: "1712800001.000300",
+      selected_option: {
+        value: "openai/gpt-5",
+      },
+    },
+    ack,
   });
 
-  test("keeps identical thread timestamps isolated by channel", () => {
-    const tracker = createAgentThreadTracker();
-    tracker.remember("C123", "1234567890.123456");
-    expect(tracker.has("C123", "1234567890.123456")).toBe(true);
-    expect(tracker.has("C999", "1234567890.123456")).toBe(false);
+  expect(ack).toHaveBeenCalledTimes(1);
+  expect(messages).toHaveLength(1);
+  expect(messages[0]).toMatchObject({
+    channel: "slack",
+    accountId: "slack-test-account",
+    chatId: "C123",
+    senderId: "U123",
+    senderName: "Alice",
+    senderTeamId: "T123",
+    text: "/model openai/gpt-5",
+    threadId: "1712800000.000200",
+    chatType: "channel",
+  });
+});
+
+test("slack adapter renders model picker blocks on direct replies", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+  const modelPicker = {
+    current: {
+      modelLabel: "Auto",
+      modelHandle: "letta/auto",
+      scope: "conversation" as const,
+    },
+    entries: [
+      {
+        id: "auto",
+        handle: "letta/auto",
+        label: "Auto",
+        description: "Recommended default",
+        isDefault: true,
+      },
+    ],
+    availableHandles: ["letta/auto"],
+    recentHandles: [],
+  };
+
+  await adapter.start();
+  await adapter.sendDirectReply(
+    "C123",
+    "Slack current conversation model: Auto.",
+    {
+      threadId: "1712800000.000200",
+      modelPicker,
+    },
+  );
+
+  const writeClient = FakeSlackWriteClient.instances[0];
+  expect(writeClient?.chat.postMessage).toHaveBeenCalledWith({
+    channel: "C123",
+    text: "Slack current conversation model: Auto.",
+    blocks: buildSlackModelPickerBlocks(modelPicker),
+    thread_ts: "1712800000.000200",
+  });
+});
+
+test("resolveSlackAccountDisplayName prefers the Slack bot profile display name", async () => {
+  await expect(
+    resolveSlackAccountDisplayName(
+      "xoxb-test-token-1234567890",
+      "xapp-test-token-1234567890",
+    ),
+  ).resolves.toBe("Letta Code (Charles Letta Code app test)");
+});
+
+test("slack adapter maps thread metadata to thread_ts", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
   });
 
-  test("clear removes all tracked threads", () => {
-    const tracker = createAgentThreadTracker();
-    tracker.remember("C123", "1111111111.111111");
-    tracker.remember("C456", "2222222222.222222");
-    expect(tracker.has("C123", "1111111111.111111")).toBe(true);
-    expect(tracker.has("C456", "2222222222.222222")).toBe(true);
-    tracker.clear();
-    expect(tracker.has("C123", "1111111111.111111")).toBe(false);
-    expect(tracker.has("C456", "2222222222.222222")).toBe(false);
+  await adapter.start();
+  await adapter.sendMessage({
+    channel: "slack",
+    chatId: "C123",
+    text: "hello",
+    threadId: "1712800000.000200",
   });
 
-  test("remember refreshes TTL for existing thread", () => {
-    let now = 0;
-    const tracker = createAgentThreadTracker({
-      maxEntries: 10,
-      now: () => now,
-      ttlMs: 100,
-    });
-    tracker.remember("C123", "1234567890.123456");
-    now = 75;
-    tracker.remember("C123", "1234567890.123456");
-    now = 125;
-    expect(tracker.has("C123", "1234567890.123456")).toBe(true);
-    now = 176;
-    expect(tracker.has("C123", "1234567890.123456")).toBe(false);
+  const writeClient = FakeSlackWriteClient.instances[0];
+  expect(writeClient).toBeDefined();
+  expect(writeClient?.options).toEqual({
+    retryConfig: {
+      retries: 0,
+    },
   });
+  expect(writeClient?.chat.postMessage).toHaveBeenCalledWith({
+    channel: "C123",
+    text: "hello",
+    thread_ts: "1712800000.000200",
+  });
+});
+
+test("slack adapter does not thread direct messages from reply metadata alone", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  await adapter.start();
+  await adapter.sendMessage({
+    channel: "slack",
+    chatId: "D123",
+    text: "hello dm",
+    replyToMessageId: "1712800000.000201",
+  });
+
+  const writeClient = FakeSlackWriteClient.instances[0];
+  expect(writeClient?.chat.postMessage).toHaveBeenCalledWith({
+    channel: "D123",
+    text: "hello dm",
+  });
+});
+
+test("slack adapter preserves explicit direct message threads", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  await adapter.start();
+  await adapter.sendMessage({
+    channel: "slack",
+    chatId: "D123",
+    text: "hello dm thread",
+    threadId: "1712800000.000200",
+    replyToMessageId: "1712800000.000201",
+  });
+
+  const writeClient = FakeSlackWriteClient.instances[0];
+  expect(writeClient?.chat.postMessage).toHaveBeenCalledWith({
+    channel: "D123",
+    text: "hello dm thread",
+    thread_ts: "1712800000.000200",
+  });
+});
+
+test("slack adapter sendDirectReply uses the dedicated write client", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  await adapter.start();
+  await adapter.sendDirectReply("C123", "reply text", {
+    replyToMessageId: "1712800000.000200",
+  });
+
+  const writeClient = FakeSlackWriteClient.instances[0];
+  expect(writeClient?.chat.postMessage).toHaveBeenCalledWith({
+    channel: "C123",
+    text: "reply text",
+    thread_ts: "1712800000.000200",
+  });
+});
+
+test("slack adapter sendDirectReply does not thread direct messages", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  await adapter.start();
+  await adapter.sendDirectReply("D123", "reply text", {
+    replyToMessageId: "1712800000.000200",
+  });
+
+  const writeClient = FakeSlackWriteClient.instances[0];
+  expect(writeClient?.chat.postMessage).toHaveBeenCalledWith({
+    channel: "D123",
+    text: "reply text",
+  });
+});
+
+test("slack adapter sendDirectReply preserves explicit direct message threads", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  await adapter.start();
+  await adapter.sendDirectReply("D123", "reply text", {
+    replyToMessageId: "1712800000.000201",
+    threadId: "1712800000.000100",
+  });
+
+  const writeClient = FakeSlackWriteClient.instances[0];
+  expect(writeClient?.chat.postMessage).toHaveBeenCalledWith({
+    channel: "D123",
+    text: "reply text",
+    thread_ts: "1712800000.000100",
+  });
+});
+
+test("slack adapter can add reactions to messages", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  await adapter.start();
+  await adapter.sendMessage({
+    channel: "slack",
+    chatId: "C123",
+    text: "",
+    reaction: ":white_check_mark:",
+    targetMessageId: "1712800000.000100",
+  });
+
+  const writeClient = FakeSlackWriteClient.instances[0];
+  expect(writeClient?.reactions.add).toHaveBeenCalledWith({
+    channel: "C123",
+    timestamp: "1712800000.000100",
+    name: "white_check_mark",
+  });
+});
+
+test("slack adapter uploads local files through Slack's external upload flow", async () => {
+  const tempDir = await mkdtemp(join(tmpdir(), "letta-slack-upload-"));
+  const mediaPath = join(tempDir, "chart.png");
+  await writeFile(mediaPath, "fake-image-data");
+
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+
+  await adapter.start();
+  const result = await adapter.sendMessage({
+    channel: "slack",
+    chatId: "C123",
+    text: "latest chart",
+    mediaPath,
+    fileName: "chart.png",
+    title: "Chart",
+    threadId: "1712790000.000050",
+  });
+
+  const writeClient = FakeSlackWriteClient.instances[0];
+  expect(writeClient?.files.getUploadURLExternal).toHaveBeenCalledWith({
+    filename: "chart.png",
+    length: "fake-image-data".length,
+  });
+  expect(fetchMock).toHaveBeenCalledWith(
+    "https://files.slack.com/upload/F123",
+    {
+      method: "POST",
+      headers: { "Content-Type": "image/png" },
+      body: expect.any(Uint8Array),
+    },
+  );
+  expect(writeClient?.files.completeUploadExternal).toHaveBeenCalledWith({
+    files: [{ id: "F123", title: "Chart" }],
+    channel_id: "C123",
+    initial_comment: "latest chart",
+    thread_ts: "1712790000.000050",
+  });
+  expect(result).toEqual({ messageId: "F123" });
+});
+
+test("slack adapter sendMessage renders a View on web context footnote when identity is provided", async () => {
+  const adapter = createSlackAdapter({
+    ...slackAccountDefaults,
+    channel: "slack",
+    enabled: true,
+    mode: "socket",
+    botToken: "xoxb-test-token-1234567890",
+    appToken: "xapp-test-token-1234567890",
+    dmPolicy: "pairing",
+    allowedUsers: [],
+  });
+  await adapter.start();
+
+  await adapter.sendMessage({
+    channel: "slack",
+    accountId: "slack-test-account",
+    chatId: "C123",
+    text: "Reply with a footnote.",
+    threadId: "1712790000.000050",
+    agentId: "agent-1",
+    conversationId: "conv-1",
+  });
+  await adapter.sendMessage({
+    channel: "slack",
+    accountId: "slack-test-account",
+    chatId: "C123",
+    text: "Reply without identity.",
+    threadId: "1712790000.000050",
+  });
+
+  const writeClient = FakeSlackWriteClient.instances[0];
+  const postCalls = writeClient?.chat.postMessage.mock
+    .calls as unknown as Array<
+    Array<{
+      channel: string;
+      text: string;
+      blocks?: Array<{
+        type: string;
+        text?: { type: string; text: string };
+        elements?: Array<{ type: string; text: string }>;
+      }>;
+    }>
+  >;
+  const withIdentity = postCalls[0]?.[0];
+  expect(withIdentity?.text).toBe("Reply with a footnote.");
+  const blocks = withIdentity?.blocks ?? [];
+  // Body rides a markdown block: section blocks clamp long text with
+  // per-block "Show more" accordions; markdown blocks render in full.
+  expect(blocks[0]).toMatchObject({
+    type: "markdown",
+    text: "Reply with a footnote.",
+  });
+  const footnoteBlock = blocks[blocks.length - 1];
+  expect(footnoteBlock?.type).toBe("context");
+  expect(footnoteBlock?.elements?.[0]?.text).toBe(
+    "<https://chat.letta.com/chat/agent-1?conversation=conv-1|View on web>",
+  );
+  // Without identity: plain text, no blocks.
+  expect(postCalls[1]?.[0]?.blocks).toBeUndefined();
 });
