+++ src/channels/slack/inbound-debounce.integration.test.ts
@@ -5,7 +5,7 @@ import {
   installSlackAdapterTestHooks,
   resolveSlackInboundAttachmentsMock,
   slackAccountDefaults,
-} from "./slack-adapter.test-harness";
+} from "./adapter-test-harness";
 
 installSlackAdapterTestHooks();
