+++ src/channels/slack/ingress-controller-events.test.ts
@@ -5,7 +5,7 @@ import {
   installSlackAdapterTestHooks,
   resolveSlackInboundAttachmentsMock,
   slackAccountDefaults,
-} from "./slack-adapter.test-harness";
+} from "./adapter-test-harness";
 
 installSlackAdapterTestHooks();
