+++ src/channels/slack/setup.ts
@@ -6,7 +6,7 @@ import {
   type DmPolicy,
   type SlackChannelAccount,
 } from "@/channels/types";
-import { resolveSlackAccountDisplayName } from "./adapter";
+import { resolveSlackAccountDisplayName } from "./account-display";
 import { ensureSlackRuntimeInstalled } from "./runtime";
 
 function isValidBotToken(token: string): boolean {
