+++ src/websocket/listener/runtime.ts
@@ -159,10 +159,7 @@ export function createConversationRuntime(
     key: runtimeKey,
     agentId: normalizedAgentId,
     conversationId: normalizedConversationId,
-    activeChannelTurnSources: null,
-    activeChannelTurnBatchId: null,
-    activeChannelTurnContextRecovered: false,
-    activeChannelTurnProgress: null,
+    activeChannelTurn: null,
     messageQueue: Promise.resolve(),
     pendingApprovalResolvers: new Map(),
     recoveredApprovalState: null,
